+++ backend/src/main/java/com/example/boongObbang/dto/TokenDto.java
@@ -1,4 +1,4 @@
-package com.example.boongObbang;
+package com.example.boongObbang.dto;
 
 import lombok.Getter;
 import lombok.Setter;
